server/requirements.txt
# Web框架
Flask>=3.0.0
Flask-CORS==3.0.10

# 云存储
qiniu==7.10.0

# HTTP请求
requests==2.28.1

# 环境变量管理
python-dotenv==0.21.0

# 开发工具
pytest==7.1.3
black==22.8.0
flake8==5.0.4